requirements.txt
flask
geopandas
pandas
numpy
shapely
rasterio
scipy
statsmodels
pillow
